Added new publish file, for Windows and Linux platforms
This one generates the binaries and archives for the following platforms:
- Windows x64
- Windows x86
- Linux x64
- Linux arm64

All platforms are compiled on Windows, and the same ConcatFiles binary is used to do the job for all platforms. Finally, they are packed into their respective archives, ready for publishing.

MacOS needs special tools since it needs to be packaged with XCode, so it cannot be included here easily